simplify throwing trees
the previous mode of operation was a bit cumbersome: think about the
file(s) one wants to expose, then think about the path one wants to map
each file to, place a ":" between the path and the filename, repeat for
all files.

this commit makes the process much easier: just specify the files (by
filename globbing, by drag&drop on windows, other means) and be done with
it. as soon as a given argument points to an existing file, it will be
exposed by it's basename. if a user needs a different mapping, he can
specify it just like before.