Feature: add -show-qr (and -serve-index)
-show-qr allows to show a QR code on the console: it is
very convenient way to scan the URL on which `knut` is
active on the phone to access whatever knut is serving.

It is only useful to provide one QR code. In a situation
when multiple resources are served, -serve-index is automatically
creating the small overview and serves it at `/`.

Also, the generated URL for the QR code requires that the
user provides a working IP: `-bind :8080` causes `knut` to
listen on all IPs on all interfaces .. but this can not be
communicated over one QR code.